Hit flash, particles, fixed errors, documentation
Added a basic hit flash and death particles ready to be tweaked on a per entity basis. Fixed all errors with loading and freeing scenes during physics processes with deferring.

Finally I wrote documentation for how to use the current system and left room for everyone to add notes on features and fixes for the game.